refactor: exception 클래스 위치 이동
