+++ service/purchasing/src/main/java/org/icd4/commerce/adapter/external/InventoryCheckerAdapter.java
@@ -2,6 +2,7 @@ package org.icd4.commerce.adapter.external;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.icd4.commerce.adapter.external.exception.ProductNotFoundException;
 import org.icd4.commerce.application.required.common.InventoryChecker;
 import org.icd4.commerce.domain.common.ProductId;
 import org.springframework.stereotype.Component;
